+++ src/main/java/de/gematik/tim/test/glue/api/fhir/practitioner/FhirSearchQuestion.java
@@ -23,9 +23,13 @@ import static de.gematik.tim.test.glue.api.utils.GlueUtils.mxidToUri;
 import static de.gematik.tim.test.glue.api.utils.RequestResponseUtils.parseResponse;
 import static de.gematik.tim.test.glue.api.utils.RequestResponseUtils.repeatedRequest;
 import static java.util.Objects.nonNull;
+import static java.util.Objects.requireNonNull;
 import static net.serenitybdd.rest.SerenityRest.lastResponse;
 
 import de.gematik.tim.test.models.FhirEndpointDTO;
+import de.gematik.tim.test.models.FhirPractitionerDTO;
+import de.gematik.tim.test.models.FhirPractitionerRoleDTO;
+import de.gematik.tim.test.models.FhirReferenceDTO;
 import de.gematik.tim.test.models.FhirResourceTypeDTO;
 import de.gematik.tim.test.models.FhirSearchResultDTO;
 import io.restassured.response.Response;
@@ -35,6 +39,7 @@ import net.serenitybdd.screenplay.Question;
 import org.apache.commons.lang3.StringUtils;
 import org.awaitility.core.ConditionTimeoutException;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -91,42 +96,68 @@ public class FhirSearchQuestion implements Question<FhirSearchResultDTO> {
   private Optional<FhirSearchResultDTO> requestFhirPractitioner(Actor actor) {
     actor.attemptsTo(SEARCH_PRACTITIONER.request().with(this::prepareQuery));
     actor.remember(LAST_RESPONSE, lastResponse());
-    if (shouldWaitTillDeleted) {
-      return dontFindPractitioner();
-    }
-    return findPractitioner();
-  }
-
-  private Optional<FhirSearchResultDTO> findPractitioner() {
     FhirSearchResultDTO response = parseResponse(FhirSearchResultDTO.class);
-    List<FhirEndpointDTO> endpoints = getResourcesFromSearchResult(response, FhirResourceTypeDTO.ENDPOINT,
-        FhirEndpointDTO.class);
-    endpoints = filterForEndpoint(endpoints);
+    List<FhirEndpointDTO> endpoints = getResourcesFromSearchResult(response, FhirResourceTypeDTO.ENDPOINT, FhirEndpointDTO.class);
+    if (nonNull(name)) {
+      List<FhirPractitionerDTO> practitioners = getResourcesFromSearchResult(response, FhirResourceTypeDTO.PRACTITIONER, FhirPractitionerDTO.class);
+      practitioners = filterByNameForEndpoint(practitioners);
+      List<String> practitionerEndpointReferences = getEndpointsForPractitioners(response, practitioners);
+      endpoints = filterByReferencesForEndpoint(practitionerEndpointReferences, endpoints);
+    }
+    if (nonNull(mxid)) {
+      endpoints = filterByMxidForEndpoint(endpoints);
+    }
+    if (shouldWaitTillDeleted) {
+      return endpoints.isEmpty() ? Optional.of(response) : Optional.empty();
+    }
     return endpoints.size() >= atLeastExpectedResults ? Optional.of(response) : Optional.empty();
   }
 
-  private Optional<FhirSearchResultDTO> dontFindPractitioner() {
-    FhirSearchResultDTO response = parseResponse(FhirSearchResultDTO.class);
-    List<FhirEndpointDTO> endpoints = getResourcesFromSearchResult(response, FhirResourceTypeDTO.ENDPOINT,
-        FhirEndpointDTO.class);
-    endpoints = filterForEndpoint(endpoints);
-    return endpoints.isEmpty() ? Optional.of(response) : Optional.empty();
+  private List<FhirEndpointDTO> filterByReferencesForEndpoint(
+      List<String> practitionerEndpointReferences,
+      List<FhirEndpointDTO> endpoints) {
+    List<FhirEndpointDTO> filteredEndpoints = new ArrayList<>();
+    for (FhirEndpointDTO endpoint : endpoints) {
+      for (String practitionerEndpointReference : practitionerEndpointReferences) {
+        if (practitionerEndpointReference.contains(requireNonNull(endpoint.getId()))) {
+          filteredEndpoints.add(endpoint);
+        }
+      }
+    }
+    return filteredEndpoints;
   }
 
-  private List<FhirEndpointDTO> filterForEndpoint(List<FhirEndpointDTO> endpoints) {
-    if (nonNull(mxid)) {
-      endpoints = endpoints.stream().filter(
-              e -> e.getAddress() != null
-                  && e.getAddress().equals(mxidToUri(mxid)))
-          .toList();
+  private List<String> getEndpointsForPractitioners(
+      FhirSearchResultDTO response,
+      List<FhirPractitionerDTO> practitioners) {
+    List<FhirPractitionerRoleDTO> practitionerRoles = getResourcesFromSearchResult(response, FhirResourceTypeDTO.PRACTITIONERROLE, FhirPractitionerRoleDTO.class);
+    List<String> endpointReferences = new ArrayList<>();
+    for (FhirPractitionerDTO practitioner : practitioners) {
+      for (FhirPractitionerRoleDTO practitionerRole : practitionerRoles) {
+        if (requireNonNull(requireNonNull(practitionerRole.getPractitioner()).getReference())
+            .contains(requireNonNull(practitioner.getId()))
+            && (nonNull(practitionerRole.getEndpoint()))) {
+          endpointReferences.addAll(practitionerRole.getEndpoint().stream().map(FhirReferenceDTO::getReference).toList());
+        }
+      }
     }
-    if (nonNull(name)) {
-      endpoints = endpoints.stream().filter(
-              e -> e.getName() != null
-                  && e.getName().contains(name))
-          .toList();
-    }
+    return endpointReferences;
+  }
+
+  private List<FhirEndpointDTO> filterByMxidForEndpoint(List<FhirEndpointDTO> endpoints) {
+    endpoints = endpoints.stream().filter(
+            endpoint -> endpoint.getAddress() != null
+                && endpoint.getAddress().equals(mxidToUri(mxid)))
+        .toList();
     return endpoints;
+  }
+
+  private List<FhirPractitionerDTO> filterByNameForEndpoint(List<FhirPractitionerDTO> practitioners) {
+    practitioners = practitioners.stream().filter(
+            practitioner -> practitioner.getName() != null
+                && requireNonNull(practitioner.getName().get(0).getText()).contains(name))
+        .toList();
+    return practitioners;
   }
 
   private RequestSpecification prepareQuery(RequestSpecification request) {